requirements-cpu.txt
# Core GUI & rendering
PySide6>=6.5.0
pyqtgraph>=0.13.0
trimesh>=3.22.0
PyOpenGL>=3.1.6
PyOpenGL-accelerate>=3.1.6
numpy>=1.23.0
pygltflib>=1.14.0

# Hunyuan3D-2mini shape-generation package
-e git+https://github.com/Tencent-Hunyuan/Hunyuan3D-2.git@main#egg=hy3dgen

# ML backends (CPU-only PyTorch build)
torch>=2.0.0
torchvision>=0.15.0
# (optional) torchaudio>=2.0.0
safetensors>=0.3.0

# Image loading
Pillow>=9.0.0

# Typing helpers
typing_extensions>=4.5.0
